# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSmart API",
    "description": "\n## 🏥 MediSmart — Hệ thống tra cứu thuốc & tư vấn sức khỏe thông minh\n\n### 🔑 Authentication\n- Lấy token: `POST /api/v1/auth/login`\n- Gắn vào header: `Authorization: Bearer {access_token}`\n- Access token hết hạn sau **30 phút** → dùng `POST /api/v1/auth/refresh`\n- Logout → token bị thu hồi ngay lập tức (Redis blacklist)\n\n### 👥 Phân quyền\n| Role  | Quyền truy cập |\n|-------|----------------|\n| User  | Hồ sơ cá nhân, tra cứu thuốc, chatbot, lịch sử |\n| Admin | Tất cả của User + quản lý dữ liệu + system logs |\n\n### 📄 Response Format\nTất cả responses đều theo format:\n```json\n{\n  \"success\": true,\n  \"message\": \"Thành công\",\n  \"data\": { ... }\n}\n```\n\n### ⚠️ Rate Limiting\n- Login: 5 lần / 15 phút / IP\n- Gửi lại email xác thực: 1 lần / 5 phút / user\n- Reset mật khẩu: 1 lần / 5 phút / user\n    ",
    "contact": {
      "name": "MediSmart Support",
      "email": "support@medismart.vn"
    },
    "license": {
      "name": "Private"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chatbot/message": {
      "post": {
        "tags": [
          "🤖 Chatbot tư vấn"
        ],
        "summary": "Gửi tin nhắn và nhận phản hồi từ AI",
        "description": "Gửi câu hỏi về sức khỏe, thuốc hoặc triệu chứng. AI sẽ trả lời dựa trên hồ sơ sức khỏe của bạn.",
        "operationId": "send_message_api_v1_chatbot_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chatbot/history": {
      "get": {
        "tags": [
          "🤖 Chatbot tư vấn"
        ],
        "summary": "Lịch sử hội thoại",
        "operationId": "get_history_api_v1_chatbot_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "🤖 Chatbot tư vấn"
        ],
        "summary": "Xóa toàn bộ lịch sử hội thoại",
        "operationId": "delete_history_api_v1_chatbot_history_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations": {
      "post": {
        "tags": [
          "✨ Gợi ý thuốc"
        ],
        "summary": "Gợi ý thuốc theo triệu chứng cá nhân hóa",
        "description": "AI phân tích triệu chứng, lịch sử bệnh và thuốc đang dùng để gợi ý thuốc phù hợp và kiểm tra tương tác.",
        "operationId": "recommend_api_v1_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/recommendations/export": {
      "post": {
        "tags": [
          "✨ Gợi ý thuốc"
        ],
        "summary": "Xuất danh sách thuốc gợi ý ra Excel",
        "operationId": "export_recommendation_api_v1_recommendations_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationResponse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessageCreate": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatMessageCreate"
      },
...
      "DrugSuggestion": {
        "properties": {
          "drug_name": {
            "type": "string",
            "title": "Drug Name"
          },
          "active_ingredient": {
            "type": "string",
            "title": "Active Ingredient"
          },
          "indication": {
            "type": "string",
            "title": "Indication"
          },
          "reference_dosage": {
            "type": "string",
            "title": "Reference Dosage"
          },
          "suitability_score": {
            "type": "integer",
            "title": "Suitability Score"
          },
          "warnings": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          },
          "drug_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drug Id"
          },
          "has_interaction": {
            "type": "boolean",
            "title": "Has Interaction",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "drug_name",
          "active_ingredient",
          "indication",
          "reference_dosage",
          "suitability_score"
        ],
        "title": "DrugSuggestion"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse": {
        "properties": {
          "items": {
            "items": {},
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "PaginatedResponse"
      },
      "PaginationMeta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "size",
          "total_pages"
        ],
        "title": "PaginationMeta"
      },
...
      "RecommendationRequest": {
        "properties": {
          "symptoms": {
            "type": "string",
            "title": "Symptoms"
          },
          "health_profile_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Profile Ids"
          },
          "current_prescription_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Prescription Id"
          }
        },
        "type": "object",
        "required": [
          "symptoms"
        ],
        "title": "RecommendationRequest"
      },
      "RecommendationResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "symptoms": {
            "type": "string",
            "title": "Symptoms"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/DrugSuggestion"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "general_advice": {
            "type": "string",
            "title": "General Advice"
          },
          "see_doctor_if": {
            "type": "string",
            "title": "See Doctor If"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "symptoms",
          "suggestions",
          "general_advice",
          "see_doctor_if",
          "generated_at"
        ],
        "title": "RecommendationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}